+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - **FEATURE:** Die alte `downloadAbos.txt` wird nun in eine sqlite3-Datenbank `abo-history.db` migriert. Die alte Textdatei bleibt danach als `downloadAbos.txt.migrated` erhalten.
 - **FEATURE:** Farben für light/dark-Mode können in den Einstellungen getrennt konfiguriert werden.
 - **FEATURE:** URL zum JDownloader kann in *Einstellungen/Erweitert* konfiguriert werden. Vielen Dank an @robNice.
+- **FEATURE:** Filme können an *pyLoad* zum Laden übergeben werden. Vielen Dank an @robNice.
 - interne Optimierungen für mehr Geschwindigkeit.
 - interne Fehler wurden behoben.